made a few small bug fixes

+++ grails-app/services/org/bbop/apollo/FeatureService.groovy
@@ -183,7 +183,7 @@ class FeatureService {
                         tmpTranscript.name = nameService.generateUniqueName(tmpTranscript, tmpGene.name)
                     }
 
-                    if (overlapperService.overlaps(tmpTranscript, tmpGene)) {
+                    if (tmpTranscript && tmpGene && overlapperService.overlaps(tmpTranscript, tmpGene)) {
                         log.debug "There is an overlap, adding to an existing gene"
                         transcript = tmpTranscript;
                         gene = tmpGene;
@@ -1988,7 +1988,7 @@ public void setTranslationEnd(Transcript transcript, int translationEnd) {
         ArrayList<Transcript> transcriptsToUpdate = new ArrayList<Transcript>()
 
         for (Transcript eachTranscript : allSortedTranscripts) {
-            if (overlapperService.overlaps(eachTranscript, fivePrimeGene)) {
+            if (eachTranscript && fivePrimeGene && overlapperService.overlaps(eachTranscript, fivePrimeGene)) {
                 if (transcriptService.getGene(eachTranscript).uniqueName != fivePrimeGene.uniqueName) {
                     transcriptsToAssociate.add(eachTranscript)
                     genesToMerge.add(transcriptService.getGene(eachTranscript))
@@ -2058,7 +2058,7 @@ public void setTranslationEnd(Transcript transcript, int translationEnd) {
                     firstTranscript.save(flush: true)
                     continue
                 }
-                if (overlapperService.overlaps(eachTranscript, firstTranscript)) {
+                if (eachTranscript && firstTranscript && overlapperService.overlaps(eachTranscript, firstTranscript)) {
                     featureRelationshipService.removeFeatureRelationship(transcriptService.getGene(eachTranscript), eachTranscript)
                     addTranscriptToGene(transcriptService.getGene(firstTranscript), eachTranscript)
                     firstTranscript.name = nameService.generateUniqueName(firstTranscript, transcriptService.getGene(firstTranscript).name)
@@ -2076,7 +2076,7 @@ public void setTranslationEnd(Transcript transcript, int translationEnd) {
         overlappingTranscripts.remove(transcript) // removing itself
         ArrayList<Transcript> transcriptsWithOverlappingOrf = new ArrayList<Transcript>()
         for (Transcript eachTranscript in overlappingTranscripts) {
-            if (overlapperService.overlaps(eachTranscript, transcript)) {
+            if (eachTranscript && transcript && overlapperService.overlaps(eachTranscript, transcript)) {
                 transcriptsWithOverlappingOrf.add(eachTranscript)
             }
         }
